Added Duration class to supplant System.currentTimeMillis().
Review by: knorton


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@2336 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/user/client/animation/WidgetAnimation.java
@@ -33,7 +33,7 @@ public abstract class WidgetAnimation extends Animation {
     protected void cancel(WidgetAnimation anim) {
     }
 
-    protected void run(WidgetAnimation anim, int duration, long startTime) {
+    protected void run(WidgetAnimation anim, int duration, double startTime) {
       anim.onRunWhenDisabled();
     }
   }
@@ -49,7 +49,7 @@ public abstract class WidgetAnimation extends Animation {
     }
 
     @Override
-    protected void run(WidgetAnimation anim, int duration, long startTime) {
+    protected void run(WidgetAnimation anim, int duration, double startTime) {
       Animation.impl.run(anim, duration, startTime);
     }
   }
@@ -82,7 +82,7 @@ public abstract class WidgetAnimation extends Animation {
    * @param startTime the synchronized start time in milliseconds
    */
   @Override
-  public void run(int duration, long startTime) {
+  public void run(int duration, double startTime) {
     widgetImpl.run(this, duration, startTime);
   }
